[feature/issue_59] - refactored builders to make children builders to return instance of its own class, instead the parent class'
